+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Option to change the title (similar to PCB Variant)
 - Render_3D: Options to disable some technical layers and control the
   silkscreen clipping. (#282)
+- Internal BoM: Now you can aggregate components using CSV files. (See #248)
 
 ### Fixed
 - Problems to compress netlists. (#287)